Remove Json based  blog import
