+++ src/JsonData/PagesData.js
@@ -26,18 +26,11 @@ const PagesData = [
         canonical_url : "https://naxrohan.github.io/my-work"
     },
     {
-        id: 4,
+        id: 5,
         title: "Blog",
         image: <RssFeedIcon />, 
         page: "/blog", 
         canonical_url : "https://naxrohan.github.io/blog"
-    },
-    {
-        id: 5,
-        title: "Posts",
-        image: <RssFeedIcon />, 
-        page: "/posts", 
-        canonical_url : "https://naxrohan.github.io/posts"
     },
 ];
