+++ src/pages/blog/[slug].jsx
@@ -1,56 +1,89 @@
-import React, { useState } from 'react'
-import Sidebar from '../../components/Sidebar'
-import { Box, createTheme, Stack, styled, ThemeProvider } from '@mui/material'
-import ResponsiveAppBar from '../../components/ResponsiveAppBar';
-import BlogArticleItem from '../../components/BlogArticleItem';
 import { useRouter } from 'next/router'
-import BlogData from "../../JsonData/BlogData"
+import React, { useState } from 'react'
+import {getPostBySlug, getAllPosts } from '../../utils/postapi'
+import markdownToHtml from '../../utils/markdownToHtml'
+import { Box, createTheme, Stack, styled, ThemeProvider } from '@mui/material'
+import ResponsiveAppBar from '../../components/ResponsiveAppBar'
+import Sidebar from '../../components/Sidebar'
+import BlogArticleItem from '../../components/BlogArticleItem'
 
-
-const BlogsPage = () => {
-
-  const router = useRouter()
-  const { slug } = router.query
-  const [mode, setMode] = useState("dark");
-  const [BlogItems, setBlogItems] = useState(BlogData.rss.channel.item);
-  // const [BlogCategories, setBlogCategories] = useState(BlogData.rss.channel.tags);
-
-  const BlogPostDetails = BlogItems.filter(item => {
-      return item.wp_post_name === slug;
-  });
-  const darkTheme = createTheme({
-    palette: {
-      mode: mode
+const BlogPost = ({post}) => {
+    const router = useRouter();
+    if(!router.isFallback && router.query.length > 0){
+      //ToDo: show 404 page
     }
-  });
-
-  const MainBody = styled(Stack)(() => ({
-      marginTop: '50px'
-  }));
-
+    const [mode, setMode] = useState("dark");
+    const darkTheme = createTheme({
+      palette: {
+        mode: mode
+      }
+    });
+  
+    const MainBody = styled(Stack)(() => ({
+        marginTop: '50px'
+    }));
   return (
     <ThemeProvider theme={darkTheme}>
-      <Box 
-        bgcolor={"background.default"} 
+      <Box
+        bgcolor={"background.default"}
         color={"text.primary"}>
-          <ResponsiveAppBar />
+        <ResponsiveAppBar />
         <Stack direction='row' spacing={3} justifyContent='space-between'>
           <Sidebar />
           <Box
             bgcolor={"background.default"}
             color={"text.primary"}
             flex={5}
-            sx={{ height: '100vh', overflowY: 'scroll', paddingLeft: { xs : 5, md: 10}, paddingRight: { xs : 5, md: 10} }}>
-              <MainBody spacing={2} justifyContent='space-evenly'>
-                <BlogArticleItem 
-                  type="full" 
-                  jsonData={BlogPostDetails[0]} />
-              </MainBody>
-        </Box>
+            sx={{ height: '100vh', overflowY: 'scroll', paddingLeft: { xs: 5, md: 10 }, paddingRight: { xs: 5, md: 10 } }}>
+            <MainBody spacing={2} justifyContent='space-evenly'>
+              <BlogArticleItem
+                type="full"
+                blogData={post} />
+            </MainBody>
+          </Box>
         </Stack>
       </Box>
     </ThemeProvider>
   )
 }
 
-export default BlogsPage
+export default BlogPost
+
+
+export async function getStaticProps({params}) {
+  const post = getPostBySlug(params.slug, [
+    'title',
+    'date',
+    'slug',
+    'author',
+    'category',
+    'content',
+    'ogImage',
+    'coverImage',
+  ])
+  const content = await markdownToHtml(post.content || '')
+  
+    return {
+      props: {
+        post: {
+          ...post,
+          content,
+        },
+      },
+    }
+}
+
+export async function getStaticPaths() {
+  const posts = getAllPosts(['slug'])
+  
+  return {
+    paths: posts.map((post) => {
+      return {
+        params: {
+          slug: post.slug,
+        },
+      }
+    }),
+    fallback: false,
+  }
+}
