+++ src/pages/blog/index.jsx
@@ -1,12 +1,13 @@
+import { Box, createTheme, Stack, ThemeProvider } from '@mui/material';
+import Link from 'next/link'
 import React, { useState } from 'react'
-import Sidebar from '../../components/Sidebar'
-import { Box, createTheme, Stack, ThemeProvider } from '@mui/material'
 import BlogListing from '../../components/BlogListing';
+import MetaTag from '../../components/MetaTags';
 import ResponsiveAppBar from '../../components/ResponsiveAppBar';
-import MetaTag from "../../components/MetaTags"
+import Sidebar from '../../components/Sidebar';
+import { getAllPosts } from '../../utils/postapi'
 
-
-const Blog = () => {
+const index = ({allPosts}) => {
   const [mode, setMode] = useState("dark");
   const darkTheme = createTheme({
     palette: {
@@ -14,23 +15,50 @@ const Blog = () => {
     }
   });
   return (
+    <ul>
+      {allPosts.map((item)=> 
+        <li>
+        <Link href={{ pathname: "/blog/[slug]", query: {slug: item.slug} }}>
+          {item.title}
+        </Link>  
+        </li>
+      )}
+    </ul>,
     <ThemeProvider theme={darkTheme}>
-      <Box 
-        bgcolor={"background.default"} 
-        color={"text.primary"}>
-          <ResponsiveAppBar />
-        <Stack direction='row' spacing={3} justifyContent='space-between'>
-          <Sidebar />
-          <BlogListing />
-        </Stack>
-        <MetaTag
-          description="naxrohan.github.io  | Blog Post Listing Page"
-          title="Blog Page"
-          siteTitle="naxrohan.github.io | ClayApps"
-          canonicalURL={`https://naxrohan.github.io/blog`} />
-      </Box>
-    </ThemeProvider>
+    <Box 
+      bgcolor={"background.default"} 
+      color={"text.primary"}>
+        <ResponsiveAppBar />
+      <Stack direction='row' spacing={3} justifyContent='space-between'>
+        <Sidebar />
+        <BlogListing
+          blogItems={allPosts} />
+      </Stack>
+      <MetaTag
+        description="naxrohan.github.io  | Blog Post Listing Page"
+        title="Blog Page"
+        siteTitle="naxrohan.github.io | ClayApps"
+        canonicalURL={`https://naxrohan.github.io/blog`} />
+    </Box>
+  </ThemeProvider>
   )
 }
 
-export default Blog
+export default index
+
+export const getStaticProps = async () => {
+  const allPosts = getAllPosts([
+    'title',
+    'date',
+    'slug',
+    'author',
+    'content',
+    'coverImage',
+    'category',
+    'excerpt',
+  ])
+
+  return {
+    props: { allPosts },
+  }
+}
